cleanup. moving types.ts to the proper dir
