+++ src/components/BaseStatistics.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import InformationSection from './InformationSection.vue';
-import type { PokemonStat } from '@/components/types';
+import type { PokemonStat } from '@/types';
 import helpers from '@/helpers';
 import { computed } from 'vue';
